requirements.txt
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
Pillow
gunicorn
